[actions] changes at multi actions
