+++ app/actions/TransactionActions.js
@@ -1,0 +1,780 @@
+import * as wallet from "wallet";
+import * as sel from "selectors";
+import eq from "lodash/fp/eq";
+import { getStakeInfoAttempt, getBalanceUpdateAttempt } from "./ClientActions";
+import { TransactionDetails } from "middleware/walletrpc/api_pb";
+import { getStartupStats } from "./StatisticsActions";
+import { hexToBytes, strHashToRaw } from "helpers";
+import { RECENT_TX_COUNT, BATCH_TX_COUNT } from "constants";
+
+export const { TRANSACTION_TYPES } = wallet;
+
+// TODO apply filters at transactions selectors
+
+// function filterTickets(tickets, filter) {
+//   return tickets
+//     .filter(v => filter.status.length ? filter.status.indexOf(v.status) > -1 : true);
+// }
+
+// export const CHANGE_TICKETS_FILTER = "CHANGE_TICKETS_FILTER";
+
+// export function changeTicketsFilter(newFilter) {
+//   return (dispatch) => {
+//     dispatch({ ticketsFilter: newFilter, type: CHANGE_TICKETS_FILTER });
+//     dispatch(getTickets());
+//   };
+// }
+
+export const GETTICKETS_CANCEL = "GETTICKETS_CANCEL";
+
+export const cancelGetTickets = () => (dispatch) =>
+  dispatch({ type: GETTICKETS_CANCEL });
+
+// filterTransactions filters a list of transactions given a filtering object.
+//
+// Currently supported filters in the filter object:
+// - type (array): Array of types a transaction must belong to, to be accepted.
+// - direction (string): A string of one of the allowed directions for regular
+//   transactions (sent/received/transferred)
+//
+// If empty, all transactions are accepted.
+function filterTransactions(transactions, filter) {
+  return transactions
+    .filter((v) =>
+      filter.types.length ? filter.types.indexOf(v.type) > -1 : true
+    )
+    .filter((v) => (filter.direction ? filter.direction === v.direction : true))
+    .filter((v) =>
+      filter.search
+        ? v.creditAddresses.find(
+            (address) =>
+              address.length > 1 &&
+              address.toLowerCase().indexOf(filter.search.toLowerCase()) !== -1
+          ) != undefined
+        : true
+    )
+    .filter((v) =>
+      filter.minAmount ? Math.abs(v.amount) >= filter.minAmount : true
+    )
+    .filter((v) =>
+      filter.maxAmount ? Math.abs(v.amount) <= filter.maxAmount : true
+    );
+}
+
+function checkAccountsToUpdate(txs, accountsToUpdate) {
+  txs.forEach((tx) => {
+    tx.tx.getCreditsList().forEach((credit) => {
+      if (accountsToUpdate.find(eq(credit.getAccount())) === undefined)
+        accountsToUpdate.push(credit.getAccount());
+    });
+    tx.tx.getDebitsList().forEach((debit) => {
+      if (accountsToUpdate.find(eq(debit.getPreviousAccount())) === undefined)
+        accountsToUpdate.push(debit.getPreviousAccount());
+    });
+  });
+  return accountsToUpdate;
+}
+
+function checkForStakeTransactions(txs) {
+  let stakeTxsFound = false;
+  txs.forEach((tx) => {
+    if (
+      tx.type == TransactionDetails.TransactionType.VOTE ||
+      tx.type == TransactionDetails.TransactionType.TICKET_PURCHASE ||
+      tx.type == TransactionDetails.TransactionType.REVOCATION
+    ) {
+      stakeTxsFound = true;
+    }
+  });
+  return stakeTxsFound;
+}
+
+export const NEW_TRANSACTIONS_RECEIVED = "NEW_TRANSACTIONS_RECEIVED";
+export const MATURINGHEIGHTS_CHANGED = "MATURINGHEIGHTS_CHANGED";
+
+// newTransactionsReceived should be called when a new set of transactions has
+// been received from the wallet (through a notification).
+export const newTransactionsReceived = (
+  newlyMinedTransactions,
+  newlyUnminedTransactions
+) => (dispatch, getState) => {
+  if (!newlyMinedTransactions.length && !newlyUnminedTransactions.length)
+    return;
+
+  let {
+    unminedTransactions,
+    recentRegularTransactions,
+    recentStakeTransactions
+  } = getState().grpc;
+  const { transactionsFilter, transactions } = getState().grpc;
+  const chainParams = sel.chainParams(getState());
+
+  // aux maps of [txhash] => tx (used to ensure no duplicate txs)
+  const newlyMinedMap = newlyMinedTransactions.reduce((m, v) => {
+    m[v.hash] = v;
+    return m;
+  }, {});
+  const newlyUnminedMap = newlyUnminedTransactions.reduce((m, v) => {
+    m[v.hash] = v;
+    return m;
+  }, {});
+
+  const unminedMap = unminedTransactions.reduce((m, v) => {
+    m[v.hash] = v;
+    return m;
+  }, {});
+
+  const unminedDupeCheck = newlyUnminedTransactions.filter(
+    (tx) => !transactions[tx.txHash] && !unminedMap[tx.txHash]
+  );
+
+  let accountsToUpdate = new Array();
+  accountsToUpdate = checkAccountsToUpdate(unminedDupeCheck, accountsToUpdate);
+  accountsToUpdate = checkAccountsToUpdate(
+    newlyMinedTransactions,
+    accountsToUpdate
+  );
+  accountsToUpdate = Array.from(new Set(accountsToUpdate));
+  accountsToUpdate.forEach((v) => dispatch(getBalanceUpdateAttempt(v, 0)));
+
+  const hasStakeTxs =
+    checkForStakeTransactions(unminedDupeCheck) ||
+    checkForStakeTransactions(newlyMinedTransactions);
+  if (hasStakeTxs) {
+    dispatch(getStakeInfoAttempt());
+  }
+
+  unminedTransactions = filterTransactions(
+    [
+      ...newlyUnminedTransactions,
+      ...unminedTransactions.filter(
+        (tx) => !newlyMinedMap[tx.hash] && !newlyUnminedMap[tx.hash]
+      )
+    ],
+    transactionsFilter
+  );
+
+  const regularTransactionFilter = {
+    listDirection: "desc",
+    types: [TransactionDetails.TransactionType.REGULAR],
+    direction: null
+  };
+
+  recentRegularTransactions = filterTransactions(
+    [
+      ...newlyUnminedTransactions,
+      ...newlyMinedTransactions,
+      ...recentRegularTransactions.filter(
+        (tx) => !newlyMinedMap[tx.hash] && !newlyUnminedMap[tx.hash]
+      )
+    ],
+    regularTransactionFilter
+  )
+    .slice(0, RECENT_TX_COUNT)
+    .map((tx) => tx);
+
+  const stakeTransactionFilter = {
+    listDirection: "desc",
+    types: [
+      TransactionDetails.TransactionType.TICKET_PURCHASE,
+      TransactionDetails.TransactionType.VOTE,
+      TransactionDetails.TransactionType.REVOCATION
+    ],
+    direction: null
+  };
+
+  recentStakeTransactions = filterTransactions(
+    [
+      ...newlyUnminedTransactions,
+      ...newlyMinedTransactions,
+      ...recentStakeTransactions.filter(
+        (tx) => !newlyMinedMap[tx.hash] && !newlyUnminedMap[tx.hash]
+      )
+    ],
+    stakeTransactionFilter
+  ).slice(0, RECENT_TX_COUNT);
+
+  const { maturingBlockHeights } = getState().grpc;
+  const newMaturingHeights = { ...maturingBlockHeights };
+  const mergeNewMaturingHeights = (hs) =>
+    Object.keys(hs).forEach((h) => {
+      const accounts = newMaturingHeights[h] || [];
+      hs[h].forEach((a) =>
+        accounts.indexOf(a) === -1 ? accounts.push(a) : null
+      );
+      newMaturingHeights[h] = accounts;
+    });
+
+  mergeNewMaturingHeights(
+    transactionsMaturingHeights(newlyMinedTransactions, chainParams)
+  );
+  dispatch({
+    maturingBlockHeights: newMaturingHeights,
+    type: MATURINGHEIGHTS_CHANGED
+  });
+
+  dispatch({
+    recentRegularTransactions,
+    recentStakeTransactions,
+    type: NEW_TRANSACTIONS_RECEIVED,
+    unminedTransactions,
+    newlyUnminedTransactions,
+    newlyMinedTransactions
+  });
+
+  if (newlyMinedTransactions.length > 0) {
+    const {
+      startupStatsEndCalcTime,
+      startupStatsCalcSeconds
+    } = getState().statistics;
+    const secFromLastStats = (new Date() - startupStatsEndCalcTime) / 1000;
+    if (secFromLastStats > 5 * startupStatsCalcSeconds) {
+      dispatch(getStartupStats());
+    }
+  }
+};
+
+export const CHANGE_TRANSACTIONS_FILTER = "CHANGE_TRANSACTIONS_FILTER";
+export function changeTransactionsFilter(newFilter) {
+  return (dispatch) => {
+    dispatch({
+      transactionsFilter: newFilter,
+      type: CHANGE_TRANSACTIONS_FILTER
+    });
+    return dispatch(getTransactions());
+  };
+}
+
+export const GETSTARTUPTRANSACTIONS_ATTEMPT = "GETSTARTUPTRANSACTIONS_ATTEMPT";
+export const GETSTARTUPTRANSACTIONS_SUCCESS = "GETSTARTUPTRANSACTIONS_SUCCESS";
+
+// getStartupTransactions fetches all recent transaction data needed for display
+// and updates. This includes:
+// - Recent regular tx data
+// - Recent stake tx data
+// - Immature future blocks (when we should update balances)
+// - Transactions since last start
+export const getStartupTransactions = () => async (dispatch, getState) => {
+  dispatch({ type: GETSTARTUPTRANSACTIONS_ATTEMPT });
+
+  const { currentBlockHeight, walletService } = getState().grpc;
+  const chainParams = sel.chainParams(getState());
+  let startRequestHeight = currentBlockHeight;
+  const pageSize = 50;
+  const voteTypes = [
+    TransactionDetails.TransactionType.VOTE,
+    TransactionDetails.TransactionType.REVOCATION
+  ];
+  const checkHeightDeltas = [
+    chainParams.TicketMaturity,
+    chainParams.CoinbaseMaturity,
+    chainParams.SStxChangeMaturity
+  ];
+  const immatureHeight = currentBlockHeight - Math.max(...checkHeightDeltas);
+
+  let foundNeededTransactions = false;
+  const recentRegularTxs = [];
+  const recentStakeTxs = [];
+  // maturingBlockHeights are kept to check when having new blocks if we need
+  // to update account's balances.
+  const maturingBlockHeights = {};
+  const votedTickets = {}; // aux map of ticket hash => true
+
+  // the mergeXXX functions pick a list of transactions returned from
+  // getTransactions and fills the appropriate result variables
+  const mergeRegularTxs = (txs) =>
+    recentRegularTxs.length < RECENT_TX_COUNT &&
+    recentRegularTxs.push(...txs.filter((tx) => !tx.isStake));
+
+  const mergeStakeTxs = (txs) =>
+    recentStakeTxs.length < RECENT_TX_COUNT &&
+    recentStakeTxs.push(
+      ...txs.filter((tx) => {
+        if (voteTypes.indexOf(tx.type) > -1) {
+          // always include vote or revocation and mark ticket as voted so we
+          // don't include it in the recent list
+          const decodedSpender = wallet.decodeRawTransaction(
+            Buffer.from(tx.tx.getTransaction()),
+            chainParams
+          );
+          const spenderInputs = decodedSpender.inputs;
+          const ticketHash = spenderInputs[spenderInputs.length - 1].prevTxId;
+          votedTickets[ticketHash] = true;
+          return true;
+        }
+        if (tx.type === TransactionDetails.TransactionType.TICKET_PURCHASE) {
+          // tickets are only added to the recent list if they haven't voted yet
+          return !votedTickets[tx.txHash];
+        }
+        return false;
+      })
+    );
+
+  const mergeImmatureHeights = (txs) => {
+    if (startRequestHeight < immatureHeight) return;
+    const txHeights = transactionsMaturingHeights(txs, chainParams);
+    Object.keys(txHeights).forEach((h) => {
+      if (h < currentBlockHeight) return;
+      const accounts = maturingBlockHeights[h] || [];
+      txHeights[h].forEach((a) =>
+        accounts.indexOf(a) === -1 ? accounts.push(a) : null
+      );
+      maturingBlockHeights[h] = accounts;
+    });
+  };
+
+  // get any unmined transactions
+  let { unmined } = await wallet.getTransactions(
+    walletService,
+    -1,
+    -1,
+    pageSize
+  );
+  unmined = await normalizeBatchTx(walletService, chainParams, unmined);
+  let transactions = unmined.reduce((m, t) => {
+    m[t.txHash] = t;
+    return m;
+  }, {});
+
+  mergeRegularTxs(unmined);
+  mergeStakeTxs(unmined);
+  mergeImmatureHeights(unmined);
+
+  while (!foundNeededTransactions) {
+    let { mined } = await wallet.getTransactions(
+      walletService,
+      startRequestHeight,
+      1,
+      pageSize
+    );
+    mined = await normalizeBatchTx(walletService, chainParams, mined);
+    // make transactions map
+    transactions = mined.reduce((m, t) => {
+      m[t.txHash] = t;
+      return m;
+    }, transactions);
+
+    if (mined.length === 0) {
+      // TODO dispatch nomoretransactions to selector
+      break; // no more transactions
+    }
+
+    mergeRegularTxs(mined);
+    mergeStakeTxs(mined);
+    mergeImmatureHeights(mined);
+
+    foundNeededTransactions =
+      recentRegularTxs.length >= RECENT_TX_COUNT &&
+      recentStakeTxs.length >= RECENT_TX_COUNT;
+
+    const lastTransaction = mined[mined.length - 1];
+    startRequestHeight = lastTransaction.height - 1;
+
+    if (startRequestHeight < 1) {
+      // TODO dispatch nomoretransactions to selector
+      break; // reached genesis
+    }
+  }
+
+  dispatch({
+    type: GETSTARTUPTRANSACTIONS_SUCCESS,
+    recentRegularTxs,
+    recentStakeTxs,
+    maturingBlockHeights,
+    transactions
+  });
+};
+
+export const DECODERAWTXS_SUCCESS = "DECODERAWTXS_SUCCESS";
+export const DECODERAWTXS_FAILED = "DECODERAWTXS_FAILED";
+
+// decodeRawTransaction requests decodification of a raw, hex-encoded transaction.
+export const decodeRawTransaction = (hexTx, hash) => (dispatch, getState) => {
+  const bufferTx = Buffer.from(hexToBytes(hexTx));
+  const chainParams = sel.chainParams(getState());
+  const decodedTx = wallet.decodeRawTransaction(bufferTx, chainParams);
+  dispatch({ decodedTx, hash, type: DECODERAWTXS_SUCCESS });
+  return decodedTx;
+};
+
+// getNonWalletOutputs decodes a tx and gets outputs which are not from the wallet.
+// This is needed to show output addresses at our home and tx history pages.
+const getNonWalletOutputs = (walletService, chainParams, tx) =>
+  new Promise((resolve, reject) => {
+    try {
+      const decodedTx = wallet.decodeRawTransaction(
+        Buffer.from(tx.getTransaction()),
+        chainParams
+      );
+      const updatedOutputs = decodedTx.outputs.map(async (o) => {
+        const address = o.decodedScript.address;
+        // Validate address so we can check if it is our own.
+        // If that is the case it is a change output.
+        const addrValidResp = await wallet.validateAddress(
+          walletService,
+          address
+        );
+        return {
+          address,
+          value: o.value,
+          isChange: addrValidResp.getIsMine()
+        };
+      });
+      resolve(Promise.all(updatedOutputs));
+    } catch (e) {
+      reject(e);
+    }
+  });
+
+export const GETTRANSACTIONS_ATTEMPT = "GETTRANSACTIONS_ATTEMPT";
+export const GETTRANSACTIONS_FAILED = "GETTRANSACTIONS_FAILED";
+export const GETTRANSACTIONS_COMPLETE = "GETTRANSACTIONS_COMPLETE";
+
+// normalizeTx is used to normalize txs after fetched from dcrwallet.
+// this is needed because we do not have all information right after
+// fetching it. So we need to manually fill them.
+const normalizeTx = async (walletService, chainParams, tx) => {
+  if (tx.isStake) {
+    // For stake txs, fetch the missing stake info.
+    return {
+      ...tx,
+      ...(await getMissingStakeTxData(walletService, chainParams, tx))
+    };
+  }
+
+  // For regular tx we get the nonWalletoutputs so we can show them at
+  // the overview.
+  const outputs = await getNonWalletOutputs(walletService, chainParams, tx.tx);
+
+  return { ...tx, outputs };
+};
+
+const normalizeBatchTx = async (walletService, chainParams, txs) =>
+  await Promise.all(
+    txs.map((tx) =>
+      (async () => await normalizeTx(walletService, chainParams, tx))()
+    )
+  );
+
+// getTransactions loads a list of transactions from the wallet, given the
+// current grpc state.
+//
+// Every call to getTransactions() tries to get BATCH_TX_COUNT
+// transactions from the wallet, given the current filter for transactions.
+// Note that more than that amount of transactions may be obtained, as the
+// iteration of transactions on the wallet is done by block height.
+//
+// When no more transactions are available given the current filter,
+// `grpc.noMoreTransactions` is set to true.
+export const getTransactions = () => async (dispatch, getState) => {
+  const {
+    currentBlockHeight,
+    getTransactionsRequestAttempt,
+    transactionsFilter,
+    walletService
+  } = getState().grpc;
+  const chainParams = sel.chainParams(getState());
+  let { noMoreTransactions, lastTransaction } = getState().grpc;
+  if (getTransactionsRequestAttempt || noMoreTransactions) return;
+
+  if (!currentBlockHeight) {
+    // Wait a little then re-dispatch this call since we have no starting height yet
+    setTimeout(() => {
+      dispatch(getTransactions());
+    }, 1000);
+    return;
+  }
+  dispatch({ type: GETTRANSACTIONS_ATTEMPT });
+
+  // first, request unmined transactions. They always come first in decrediton.
+  let { unmined } = await wallet.getTransactions(walletService, -1, -1, 0);
+  unmined = await normalizeBatchTx(walletService, chainParams, unmined);
+
+  // TODO: Filter and get other transactions data in another stage. Use getTransactions only
+  // to get transactions and dispatch them at transactions's map.
+
+  // now, request a batch of mined transactions until BATCH_TX_COUNT
+  // transactions have been obtained (after filtering)
+  const reachedGenesis = false;
+  let transactions = unmined.reduce((m, t) => {
+    m[t.txHash] = t;
+    return m;
+  }, {});
+
+  while (!noMoreTransactions && !reachedGenesis) {
+    let startRequestHeight, endRequestHeight;
+    const transactionsLength = Object.keys(transactions).length;
+
+    if (transactionsLength >= BATCH_TX_COUNT) break;
+
+    if (transactionsFilter.listDirection === "desc") {
+      endRequestHeight = 1;
+      startRequestHeight = lastTransaction
+        ? lastTransaction.height - 1
+        : currentBlockHeight;
+      // Reached genesis.
+      if (startRequestHeight < 1) {
+        noMoreTransactions = true;
+        break;
+      }
+    } else {
+      startRequestHeight = lastTransaction ? lastTransaction.height + 1 : 1;
+      endRequestHeight = currentBlockHeight;
+    }
+
+    try {
+      let { mined } = await wallet.getTransactions(
+        walletService,
+        startRequestHeight,
+        endRequestHeight,
+        BATCH_TX_COUNT
+      );
+      // no more transactions
+      if (mined.length === 0) {
+        noMoreTransactions = true;
+        break;
+      }
+      lastTransaction = mined[mined.length - 1];
+
+      mined = await normalizeBatchTx(walletService, chainParams, mined);
+
+      // make transactions map
+      transactions = mined.reduce((m, t) => {
+        m[t.txHash] = t;
+        return m;
+      }, transactions);
+    } catch (error) {
+      dispatch({ type: GETTRANSACTIONS_FAILED, error });
+      return;
+    }
+  }
+
+  return dispatch({
+    type: GETTRANSACTIONS_COMPLETE,
+    noMoreTransactions,
+    lastTransaction,
+    transactions
+  });
+};
+
+const getMissingStakeTxData = async (
+  walletService,
+  chainParams,
+  transaction
+) => {
+  let ticketTx, spenderTx, status;
+  const { txHash, txType, rawTx, tx } = transaction;
+
+  if (txType === "Ticket") {
+    // This is currently a somewhat slow call in RPC mode due to having to check
+    // in dcrd whether the ticket is live or not.
+    const ticket = await wallet.getTicket(walletService, strHashToRaw(txHash));
+    status = ticket.status;
+    if (ticket.spender.getHash()) {
+      try {
+        const spenderHash = ticket.spender.getHash();
+        const spender = await wallet.getTransaction(walletService, spenderHash);
+        spenderTx = spender.tx;
+      } catch (error) {
+        if (String(error).indexOf("NOT_FOUND") === -1) {
+          // A NOT_FOUND error means the wallet hasn't recorded the spender of
+          // this ticket (possibly it was purchased with a reward address from
+          // a different wallet). So just return the available information in
+          // that case and error out in actual failures.
+          throw error;
+        }
+      }
+    }
+    ticketTx = tx;
+  } else {
+    // vote/revoke
+    const decodedSpender = wallet.decodeRawTransaction(
+      Buffer.from(rawTx, "hex"),
+      chainParams
+    );
+
+    // Find the ticket hash of the vote/revoke.
+    const spenderInputs = decodedSpender.inputs;
+    const outpIdx = spenderInputs.length === 1 ? 0 : 1;
+    const ticketTxHash = spenderInputs[outpIdx].prevTxId;
+    status = txType === "Vote" ? "voted" : "revoked";
+
+    // given that the ticket may be much older than the transactions currently
+    // in the `transactions` state var, we need to manually fetch the ticket
+    // transaction
+    try {
+      const ticket = await wallet.getTransaction(walletService, ticketTxHash);
+      ticketTx = ticket.tx;
+    } catch (error) {
+      if (String(error).indexOf("NOT_FOUND") > -1) {
+        // NOT_FOUND error means we have a vote/revocation tx recorded but not
+        // the originating ticket. This might happen in pool fee wallets or
+        // wallets receiving the rewards from a different funding wallet.
+        // In that case, we return with the original tx since there's nothing
+        // else to do.
+      } else {
+        throw error;
+      }
+    }
+    spenderTx = tx;
+  }
+
+  return { ticket: ticketTx, spender: spenderTx, status };
+};
+
+export const FETCHMISSINGSTAKETXDATA_ATTEMPT =
+  "FETCHMISSINGSTAKETXDATA_ATTEMPT";
+export const FETCHMISSINGSTAKETXDATA_SUCCESS =
+  "FETCHMISSINGSTAKETXDATA_SUCCESS";
+export const FETCHMISSINGSTAKETXDATA_FAILED = "FETCHMISSINGSTAKETXDATA_FAILED";
+
+// fetchMissingStakeTxData fetches the missing stake information of a given
+// transaction returned from getTransaction. For tickets, it tries to fill the
+// ticket purchase amount and for votes it tries to fill the purchase info and
+// reward data.
+//
+// This function is not suitable for calling on a list of transactions, since
+// some cases/operations are slow.
+export const fetchMissingStakeTxData = (tx) => async (dispatch, getState) => {
+  const { walletService, fetchMissingStakeTxDataAttempt } = getState().grpc;
+  const chainParams = sel.chainParams(getState());
+  if (fetchMissingStakeTxDataAttempt[tx.txHash]) {
+    return;
+  }
+  dispatch({ txHash: tx.txHash, type: FETCHMISSINGSTAKETXDATA_ATTEMPT });
+
+  const newTx = await getMissingStakeTxData(walletService, chainParams, tx);
+
+  const oldTxs = getState().grpc.transactions;
+  const oldStakeTxs = getState().grpc.recentStakeTransactions;
+  const txIdx = oldTxs.findIndex((t) => t.txHash === tx.txHash);
+  const stakeTxIdx = oldStakeTxs.findIndex((t) => t.txHash === tx.txHash);
+
+  if (txIdx > -1 || stakeTxIdx > -1) {
+    const dispatchState = {
+      txHash: tx.txHash,
+      type: FETCHMISSINGSTAKETXDATA_SUCCESS
+    };
+    if (txIdx > -1) {
+      const newTxs = [...oldTxs];
+      newTxs[txIdx] = newTx;
+      dispatchState["transactions"] = newTxs;
+    }
+    if (stakeTxIdx > -1) {
+      const newStakeTxs = [...oldStakeTxs];
+      newStakeTxs[stakeTxIdx] = newTx;
+      dispatchState["recentStakeTransactions"] = newStakeTxs;
+    }
+    dispatch(dispatchState);
+  } else {
+    // not supposed to happen in normal usage; this function  should only be
+    // entered from a transaction already in the transaction map
+    // (sel.transactionsMap).
+    dispatch({ txHash: tx.txHash, type: FETCHMISSINGSTAKETXDATA_FAILED });
+  }
+};
+
+// Given a list of transactions, returns the maturing heights of all
+// stake txs in the list.
+function transactionsMaturingHeights(txs, chainParams) {
+  const res = {};
+  const addToRes = (height, found) => {
+    const accounts = res[height] || [];
+    found.forEach((a) =>
+      accounts.indexOf(a) === -1 ? accounts.push(a) : null
+    );
+    res[height] = accounts;
+  };
+
+  txs.forEach((tx) => {
+    const accountsToUpdate = [];
+    switch (tx.type) {
+      case TransactionDetails.TransactionType.TICKET_PURCHASE:
+        checkAccountsToUpdate([tx], accountsToUpdate);
+        addToRes(tx.height + chainParams.TicketExpiry, accountsToUpdate);
+        addToRes(tx.height + chainParams.SStxChangeMaturity, accountsToUpdate);
+        addToRes(tx.height + chainParams.TicketMaturity, accountsToUpdate); // FIXME: remove as it doesn't change balances
+        break;
+
+      case TransactionDetails.TransactionType.VOTE:
+      case TransactionDetails.TransactionType.REVOCATION:
+        checkAccountsToUpdate([tx], accountsToUpdate);
+        addToRes(tx.height + chainParams.CoinbaseMaturity, accountsToUpdate);
+        break;
+    }
+  });
+
+  return res;
+}
+
+// getAmountFromTxInputs receives a decoded unsignedTx and adds the amount of
+// each input from the previous transaction. We need this because when decoding
+// a tx, the amount of the input is not deoded with it.
+export const getAmountFromTxInputs = (unsignedTx) => async (
+  dispatch,
+  getState
+) => {
+  const { walletService } = getState().grpc;
+  const chainParams = sel.chainParams(getState());
+  const inputWithAmount = unsignedTx.inputs.map(
+    async ({ prevTxId, outputIndex }, i) => {
+      const oldTx = await wallet.getTransaction(walletService, prevTxId);
+      if (!oldTx) {
+        return Promise.reject(new Error(`Transaction ${prevTxId} not found`));
+      }
+      const rawTxBuffer = Buffer.from(hexToBytes(oldTx.rawTx));
+      const decodedOldTx = wallet.decodeRawTransaction(
+        rawTxBuffer,
+        chainParams
+      );
+      const oldOutput = decodedOldTx.outputs[outputIndex];
+      if (!oldOutput) {
+        return Promise.reject(
+          new Error(`Unknown outpoint ${prevTxId}:${outputIndex} not found`)
+        );
+      }
+      unsignedTx.inputs[i].amountIn = decodedOldTx.outputs[outputIndex].value;
+      // we also save the outpoint address to perform sanity checkes on trezor txs.
+      unsignedTx.inputs[i].outpointAddress =
+        decodedOldTx.outputs[outputIndex].addresses[0].address;
+      unsignedTx.inputs[i].outpoint = `${prevTxId}:${outputIndex}`;
+      return unsignedTx.inputs[i];
+    }
+  );
+
+  unsignedTx.inputs = await Promise.all(inputWithAmount);
+
+  return unsignedTx;
+};
+
+// getTxFromInputs receives a decoded unsignedTx as parameter and gets the txs
+// which participate in the input.
+export const getTxFromInputs = (unsignedTx) => (dispatch, getState) =>
+  new Promise((resolve, reject) => {
+    const { walletService } = getState().grpc;
+    const chainParams = sel.chainParams(getState());
+    // aux map to know if we already has gotten a tx.
+    const txsMap = {};
+    const txs = unsignedTx.inputs.reduce(async (txs, inp) => {
+      const { prevTxId } = inp;
+      if (txsMap[prevTxId]) return;
+
+      txsMap[prevTxId] = true;
+      const oldTx = await wallet.getTransaction(walletService, prevTxId);
+      if (!oldTx) {
+        return reject(new Error(`Transaction ${prevTxId} not found`));
+      }
+
+      const rawTxBuffer = Buffer.from(hexToBytes(oldTx.rawTx));
+      const decodedOldTx = wallet.decodeRawTransaction(
+        rawTxBuffer,
+        chainParams
+      );
+
+      txs.push(decodedOldTx);
+
+      return txs;
+    }, []);
+
+    resolve(txs);
+  });
